Link SceneSettings Popout Menu
Scene Settings popout menu is now linked to the scene and saves the prefs out to a file for loading when the scene loads. Sprites are the only object that would probably need to use this default object size so may change the text on the gui to reflect that.