Small Update
This small update fixes for future settings for update callbacks. This makes it so if you add a behavior to an object that is based on time such as a spawner as soon as you press the play button it will start the scene at 0 time and then the pause will revert the scene back to a state it was at before the play button was pressed. Pause button is in effect is now a stop button so i will change the gui art to reflect this soon. This is ground work for update callbacks and scene timing for gameplay.